Update qmedianbuffer.h
changed name of function to avoid conflict; some small changes, and quick sort algorithm is still in, but inactive. it is to be here if someone is to try it out in real world if quicksort is faster for finding median, or max and min values. In that case, sequence is: order numbers, reset oldest to zero, do quicksort (now. most numbers will change positions, but maybe not all). later, sorting is needed before push. If quicksort is faster, it should be used for three things only - median (the one that returns only original value), minValue and maxValue (and equivalent rate of change functions).